add lit-html as render function
