+++ demo/index.html
@@ -8,6 +8,7 @@
 </head>
 <body>
   <script src="index.js"></script>  
-  <demo-component data-demo="Current time:"></demo-component>
+  <demo-hyper-component data-demo="Current time is rendering with hyperHTML:"></demo-hyper-component>
+  <demo-lit-component data-demo="Current time is rendering with lit-html:"></demo-lit-component>
 </body>
 </html>
