+++ demo/index.ts
@@ -1,9 +1,10 @@
-import { AbstractElement, Define } from '../dist/abstract-element';
+import { Define, AbstractLitElement, AbstractHyperElement } from '../dist/abstract-element';
+import { TemplateResult } from 'lit-html';
 
 
 
-@Define('demo-component')
-export class DemoComponent extends AbstractElement {
+@Define('demo-hyper-component')
+export class DemoHyperComponent extends AbstractHyperElement {
   static attrNames = { dataDemo: 'data-demo' };
   state = {
     time: new Date().toLocaleTimeString()
@@ -11,7 +12,7 @@ export class DemoComponent extends AbstractElement {
 
 
   constructor() {
-    super();
+    super('', true);
 
     // update the time each second
     setInterval(() => {
@@ -23,6 +24,38 @@ export class DemoComponent extends AbstractElement {
 
 
   render() {
+    return this.html`
+    <section>
+      <p>${this.attr[DemoComponent.attrNames.dataDemo]}</p>
+      <div>${this.state.time}</div>
+    </section>
+    `;
+  }
+}
+
+
+
+@Define('demo-lit-component')
+export class DemoComponent extends AbstractLitElement {
+  static attrNames = { dataDemo: 'data-demo' };
+  state = {
+    time: new Date().toLocaleTimeString()
+  };
+
+
+  constructor() {
+    super('', true);
+
+    // update the time each second
+    setInterval(() => {
+      this.state = {
+        time: new Date().toLocaleTimeString()
+      };
+    }, 1000);
+  }
+
+
+  render(): TemplateResult {
     return this.html`
     <section>
       <p>${this.attr[DemoComponent.attrNames.dataDemo]}</p>
